Create Occurrence of a Character
Program Description
Create a program to print the occurrence of a character in the string.

Take string input and store it in the string1 variable.
Take a character input (a string with a single character) and store it in character1 variable.
Create a variable named count and initialize it to 0.
Use a for loop to iterate through string1.
Inside the loop, check the occurrence of the character in the string. If it's True, increment count by 1.
Print the count variable.
Assumption: We will assume the user will first enter a string and then a single character.